backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
httpx==0.28.1
openai==1.57.2
pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
python-jose==3.3.0
